fix: process review comments
