+++ examples/CMakeLists.txt
@@ -2,7 +2,6 @@ if (EMIL_HOST_BUILD)
     add_subdirectory(http)
     add_subdirectory(https)
     add_subdirectory(mdns)
-    add_subdirectory(network_example)
     add_subdirectory(serial_net)
     add_subdirectory(sntp)
 endif()
