+++ examples/https/CMakeLists.txt
@@ -1,3 +1,3 @@
 add_executable(examples.https_client)
 target_sources(examples.https_client PRIVATE Client.cpp)
-target_link_libraries(examples.https_client PRIVATE examples.network_example hal.generic)
+target_link_libraries(examples.https_client PRIVATE services.network_instantiations hal.generic)
