+++ examples/mdns/Client.cpp
@@ -1,13 +1,14 @@
-#include "examples/network_example/Network.hpp"
 #include "hal/generic/TimerServiceGeneric.hpp"
 #include "infra/timer/Timer.hpp"
 #include "services/network/MdnsClient.hpp"
+#include "services/network_instantiations/NetworkAdapter.hpp"
 #include "services/tracer/TracerOnIoOutputInfrastructure.hpp"
 
-extern "C" int Main(main_::Network& network, int argc, const char* argv[], const char* env[])
+int main(int argc, const char* argv[], const char* env[])
 {
     static hal::TimerServiceGeneric timerService;
     static main_::TracerOnIoOutputInfrastructure tracer;
+    static main_::NetworkAdapter network;
     static services::MdnsClient mdns(network.DatagramFactory(), network.Multicast(), services::IPVersions::ipv4);
     static services::MdnsQueryImpl query(mdns, services::DnsType::dnsTypeA, "discovery", [](infra::ConstByteRange data)
     {
