+++ services/network_bsd/CMakeLists.txt
@@ -11,8 +11,8 @@ if (EMIL_BUILD_UNIX OR EMIL_BUILD_DARWIN)
         DatagramBsd.hpp
         EventDispatcherWithNetwork.cpp
         EventDispatcherWithNetwork.hpp
-        NameLookupBsd.cpp
-        NameLookupBsd.hpp
+        NameLookup.cpp
+        NameLookup.hpp
     )
 
     emil_clangformat_target(services.network_bsd)
