+++ services/network_bsd/NameLookup.cpp
@@ -1,15 +1,17 @@
+#include "services/network_bsd/NameLookup.hpp"
 #include "infra/event/EventDispatcher.hpp"
-#include "services/network_win/NameLookupWin.hpp"
-#include <winsock2.h>
-#include <ws2tcpip.h>
+#include <netdb.h>
+#include <sys/socket.h>
+#include <sys/types.h>
 
 namespace services
 {
-    NameLookupWin::NameLookupWin()
-        : lookupThread([this]() { Run(); })
+    NameLookup::NameLookup()
+        : lookupThread([this]()
+              { Run(); })
     {}
 
-    NameLookupWin::~NameLookupWin()
+    NameLookup::~NameLookup()
     {
         {
             std::lock_guard<std::mutex> lock(mutex);
@@ -21,7 +23,7 @@ namespace services
         lookupThread.join();
     }
 
-    void NameLookupWin::Lookup(NameResolverResult& result)
+    void NameLookup::Lookup(NameResolverResult& result)
     {
         assert(result.Versions() != IPVersions::ipv6);
         std::lock_guard<std::mutex> lock(mutex);
@@ -29,13 +31,13 @@ namespace services
         condition.notify_one();
     }
 
-    void NameLookupWin::CancelLookup(NameResolverResult& result)
+    void NameLookup::CancelLookup(NameResolverResult& result)
     {
         std::lock_guard<std::mutex> lock(mutex);
         nameLookup.erase(result);
     }
 
-    void NameLookupWin::Run()
+    void NameLookup::Run()
     {
         std::unique_lock<std::mutex> lock(mutex);
 
@@ -55,7 +57,7 @@ namespace services
         }
     }
 
-    void NameLookupWin::LookupIPv4()
+    void NameLookup::LookupIPv4()
     {
         std::string terminatedHostname{ currentLookup->Hostname().begin(), currentLookup->Hostname().end() };
         terminatedHostname.push_back('\0');
@@ -67,22 +69,21 @@ namespace services
                 if (entry->ai_family == AF_INET)
                 {
                     sockaddr_in* address = reinterpret_cast<sockaddr_in*>(entry->ai_addr);
-                    auto ipv4Address = services::IPv4Address{ address->sin_addr.s_net, address->sin_addr.s_host, address->sin_addr.s_lh, address->sin_addr.s_impno };
+                    auto ipv4Address = services::ConvertFromUint32(ntohl(address->sin_addr.s_addr));
                     infra::EventDispatcher::Instance().Schedule([this, ipv4Address]()
-                    {
+                        {
                         std::lock_guard<std::mutex> lock(mutex);
                         if (&nameLookup.front() == currentLookup)
                         {
                             nameLookup.pop_front();
                             currentLookup->NameLookupDone(ipv4Address, infra::Now() + std::chrono::minutes(5));
                         }
-                        condition.notify_one();
-                    });
+                        condition.notify_one(); });
                     return;
                 }
 
         infra::EventDispatcher::Instance().Schedule([this]()
-        {
+            {
             std::unique_lock<std::mutex> lock(mutex);
             if (&nameLookup.front() == currentLookup)
             {
@@ -91,7 +92,6 @@ namespace services
                 lock.unlock();
                 lookup->NameLookupFailed();
             }
-            condition.notify_one();
-        });
+            condition.notify_one(); });
     }
 }
